Begin conversion to new spider
