--- Alternate Code/useragent
@@ -1,1 +1,0 @@
-"Mozilla/5.0 (X11; Ubuntu; Linux i686; rv:48.0) Gecko/20100101 Firefox/48.0"
